Rename `azul.dos` to `azul.drs`
